requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-filter==24.2
django-prometheus==2.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
Markdown==3.6
multidict==6.0.5
packaging==24.0
prometheus_client==0.20.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.5.1
sqlparse==0.5.0
twilio==9.0.5
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.0.0
yarl==1.9.4
